Use IThreadPoolWorkItem in ConcurrentExclusiveSchedulerPair (#20513)
ConcurrentExclusiveSchedulerPair exposes two TaskSchedulers, one for concurrent processing and one for exclusive processing, such that any tasks scheduled are executed in a corresponding fashion.  CESP does that by itself scheduling workers to an underlying scheduler, and those workers in turn process the queued tasks.  Today this always queues those workers as Tasks, but if the underlying scheduler is TaskScheduler.Default (aka the thread pool), we can avoid per-worker allocations by using the new UnsafeQueueUserWorkItem overload that takes an IThreadPoolWorkItem, and just queueing the same worker object over and over and over.